Add out-of-tree build support
configure has a new option "--outdir path/to/dir" which will allow all build files
to be generated outside of the source directory (config.mk and config.gyi are
still created in the source directory).

Files will be generated in "path/to/dir/out" and the symlink to the executables
will be created at "path/to/dir/node{,_g}".

Signed-off-by: TJ <tj@khiasmos.org>